+++ src/Concerto/TestBundle/Controller/TestRunnerController.php
@@ -3,6 +3,7 @@
 namespace Concerto\TestBundle\Controller;
 
 use Concerto\TestBundle\Service\TestRunnerService;
+use Symfony\Component\HttpFoundation\Cookie;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use Psr\Log\LoggerInterface;
@@ -119,6 +120,13 @@ class TestRunnerController
         );
         $response = new Response($result);
         $response->headers->set('Content-Type', 'application/json');
+
+        $decodedResult = json_decode($result, true);
+        $cookies = $decodedResult["data"]["cookies"];
+        foreach ($cookies as $k => $v) {
+            $response->headers->setCookie(new Cookie($k, $v));
+        }
+
         return $response;
     }
 
@@ -154,6 +162,13 @@ class TestRunnerController
         );
         $response = new Response($result);
         $response->headers->set('Content-Type', 'application/json');
+
+        $decodedResult = json_decode($result, true);
+        $cookies = $decodedResult["data"]["cookies"];
+        foreach ($cookies as $k => $v) {
+            $response->headers->setCookie(new Cookie($k, $v));
+        }
+
         return $response;
     }
 
@@ -187,6 +202,13 @@ class TestRunnerController
         );
         $response = new Response($result);
         $response->headers->set('Content-Type', 'application/json');
+
+        $decodedResult = json_decode($result, true);
+        $cookies = $decodedResult["data"]["cookies"];
+        foreach ($cookies as $k => $v) {
+            $response->headers->setCookie(new Cookie($k, $v));
+        }
+
         return $response;
     }
 
@@ -209,6 +231,12 @@ class TestRunnerController
         );
         $response = new Response($result);
         $response->headers->set('Content-Type', 'application/json');
+
+        $decodedResult = json_decode($result, true);
+        $cookies = $decodedResult["data"]["cookies"];
+        foreach ($cookies as $k => $v) {
+            $response->headers->setCookie(new Cookie($k, $v));
+        }
 
         return $response;
     }